+++ docs/changelog.rst
@@ -12,7 +12,7 @@ Fixed
 Changed
 ~~~~~~~
 
--  Add support for Python 3.13.
+-  Add support for Python 3.13, 3.14.
 -  Drop support for Python 3.9.
 
 0.3.0 (2024-09-15)