+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Fixed
 ~~~~~
 
 -  Use the new ``ArrayItemsWarning`` warning, instead of the incorrect ``DeepPropertiesWarning`` warning, in :meth:`jscc.testing.checks.validate_array_items`.
+-  Break cycles in recursive schemas.
 
 Removed
 ~~~~~~~~